qd2_bootstrap_cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qd2_bootstrap"
version = "0.1.0"
description = "CLI para bootstrap de despliegues de Quditto en Kubernetes"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "rich>=13.0",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "kubernetes>=29.0.0"
]

[project.scripts]
qd2_bootstrap = "qd2_bootstrap.cli:run"

[tool.setuptools.packages.find]
include = ["qd2_bootstrap*"]
